feat(#1397): CLEAN: Split 307-line handlePostAgentSuccess in stages.ts
